+++ crates/rsvelte_core/src/svelte2tsx/template/mod.rs
@@ -15,9 +15,9 @@ mod utils;
 mod walk;
 
 use crate::ast::template::{
-    Attribute, AttributeValue, AttributeValuePart, AwaitBlock, Component, EachBlock, Fragment,
-    IfBlock, KeyBlock, LetDirective, RegularElement, SlotElement, SnippetBlock,
-    SvelteComponentElement, SvelteDynamicElement, SvelteElement, TemplateNode, TitleElement,
+    Attribute, AttributeValue, AttributeValuePart, Component, Fragment, LetDirective,
+    RegularElement, SlotElement, SvelteComponentElement, SvelteDynamicElement, SvelteElement,
+    TemplateNode, TitleElement,
 };
 use std::fmt::Write as _;
 
@@ -61,6 +61,7 @@ use utils::source::{
 
 pub(crate) use ctx::{clear_element_opener_comments, set_element_opener_comments};
 use nodes::attach_tag::format_attach_tag_segments;
+use nodes::snippet_block::{handle_snippet_block_as_component_prop, hoist_snippet_blocks};
 use walk::{process_fragment_inplace, process_node_inplace};
 
 // =============================================================================
@@ -141,57 +142,6 @@ pub fn collect_template_info(fragment: &Fragment, source: &str) -> TemplateInfo 
     info
 }
 
-/// Hoist `{#snippet}` blocks to the top of their containing block/element.
-///
-/// Mirrors `hoistSnippetBlock` in the JS reference
-/// (`htmlxtojsx_v2/nodes/SnippetBlock.ts`): each non-leading snippet child is
-/// moved to `targetPosition`, the position of the first non-snippet,
-/// non-empty-text child. This lets later content reference a snippet defined
-/// further down in source (the generated `const foo = ...` declaration is
-/// emitted before the `{const}` / `{let}` declaration tags and elements that
-/// follow it).
-///
-/// Snippets that are already first (`targetPosition` still `None`) or already
-/// at the target position are left untouched, matching the JS reference's
-/// early-`continue` guards. Component / boundary containers are excluded by
-/// their callers (they treat snippets as implicit props instead), so this is
-/// only invoked for block and plain-element fragments.
-fn hoist_snippet_blocks(fragment: &Fragment, source: &str, str: &mut MagicString) {
-    let mut target_position: Option<u32> = None;
-    for node in &fragment.nodes {
-        if !matches!(node, TemplateNode::SnippetBlock(_)) {
-            if target_position.is_none() {
-                let is_empty_text = match node {
-                    TemplateNode::Text(t) => source
-                        .get(t.start as usize..t.end as usize)
-                        .map(|s| s.trim().is_empty())
-                        .unwrap_or(true),
-                    _ => false,
-                };
-                if !is_empty_text {
-                    // JS reference: `node.type === 'Text' ? node.end : node.start`
-                    target_position = Some(match node {
-                        TemplateNode::Text(t) => t.end,
-                        _ => node.start(),
-                    });
-                }
-            }
-            continue;
-        }
-
-        // It's a snippet block.
-        let Some(tp) = target_position else {
-            // Already the first meaningful child — nothing to move.
-            continue;
-        };
-        let s = node.start();
-        if s == tp {
-            continue;
-        }
-        str.move_range(s, node.end(), tp);
-    }
-}
-
 // =============================================================================
 // Text and Comments
 // =============================================================================
@@ -203,1150 +153,6 @@ fn hoist_snippet_blocks(fragment: &Fragment, source: &str, str: &mut MagicString
 // =============================================================================
 // Block Nodes
 // =============================================================================
-
-/// Handle an if block: `{#if condition}...{:else if}...{:else}...{/if}`.
-///
-/// Generates: `if(show){...} else {...}`
-fn handle_if_block(
-    block: &IfBlock,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    if block.start >= block.end {
-        return;
-    }
-
-    let test_text = get_expression_text(&block.test, source);
-
-    // Find the start of the consequent content. When the consequent is empty
-    // (`{#if x}{:else …}` / `{#if x}{/if}`), the body still opens right after
-    // the `}` that closes the `{#if EXPR}` (or `{:else if EXPR}`) tag — this
-    // mirrors official `handleIf`, which always places the `){` body opener at
-    // `indexOf('}', expressionEnd) + 1`. Using `block.end` here (the position
-    // after `{/if}`) made the header overwrite swallow the entire `{:else …}`
-    // / `{/if}` tail, corrupting the output.
-    let consequent_start = if !block.consequent.nodes.is_empty() {
-        block.consequent.nodes[0].start()
-    } else {
-        let test_end = get_expression_range(&block.test)
-            .map(|(_, e)| e)
-            .unwrap_or(block.start);
-        let bytes = source.as_bytes();
-        let mut p = test_end as usize;
-        while p < bytes.len() && bytes[p] != b'}' {
-            p += 1;
-        }
-        ((p + 1).min(bytes.len())) as u32
-    };
-
-    // Mirror `htmlxtojsx_v2/nodes/IfElseBlock.ts::handleIf`: an IfBlock that
-    // is the elseif branch of an outer IfBlock starts at the `{` of
-    // `{:else if EXPR}` (with `expression.start` *before* `block.start` —
-    // svelte 5 records the test expression at its source-level position).
-    // Overwrite `{:else if ` → `} else if (` and the trailing `}` → `){`,
-    // exactly as the JS reference does.
-    if block.elseif {
-        let (test_start, test_end) = get_expression_range(&block.test).unwrap_or((0, 0));
-        let bytes = source.as_bytes();
-        let mut brace_open = test_start as usize;
-        while brace_open > 0 && bytes[brace_open - 1] != b'{' {
-            brace_open -= 1;
-        }
-        brace_open = brace_open.saturating_sub(1);
-        str.overwrite(brace_open as u32, test_start, "} else if (");
-
-        let mut close_brace = test_end as usize;
-        while close_brace < bytes.len() && bytes[close_brace] != b'}' {
-            close_brace += 1;
-        }
-        if close_brace < bytes.len() {
-            str.overwrite(test_end, (close_brace + 1) as u32, "){");
-        }
-    } else {
-        // Split the `{#if EXPR}` rewrite so the test expression stays as
-        // an unchanged source chunk in MagicString — preserves
-        // per-character source-map segments for TS diagnostics inside
-        // the condition. Falls back to the bulk `overwrite` when the
-        // expression has no concrete source range (e.g. synthesised).
-        if let Some((test_start, test_end)) = get_expression_range(&block.test)
-            && test_start >= block.start
-            && test_end <= consequent_start
-        {
-            str.overwrite(block.start, test_start, "if(");
-            // [test_start, test_end) left untouched.
-            if test_end < consequent_start {
-                str.overwrite(test_end, consequent_start, ")");
-            } else {
-                str.append_left(consequent_start, ")");
-            }
-        } else {
-            str.overwrite(block.start, consequent_start, &format!("if({})", test_text));
-        }
-        // Insert opening brace
-        str.append_left(consequent_start, "{");
-    }
-
-    // Hoist inner snippets above sibling `{@const}`/`{let}` / elements that
-    // reference them (a `{@const xx = test}` before its `{#snippet test}` in the
-    // same block needs `test` declared first), as in the each-body path.
-    hoist_snippet_blocks(&block.consequent, source, str);
-
-    // Process children (blocks don't increment depth)
-    process_fragment_inplace(&block.consequent, source, options, str, counter, depth);
-
-    // Handle alternate
-    if let Some(ref alternate) = block.alternate {
-        hoist_snippet_blocks(alternate, source, str);
-        // Find the {:else} or {:else if} tag position
-        // The alternate fragment starts after the {:else} tag
-        let alternate_start = if !alternate.nodes.is_empty() {
-            alternate.nodes[0].start()
-        } else {
-            block.end
-        };
-
-        // Check if the alternate is an elseif
-        let has_elseif =
-            alternate.nodes.len() == 1 && matches!(alternate.nodes[0], TemplateNode::IfBlock(_));
-
-        if has_elseif {
-            // Don't insert anything between consequent end and the nested
-            // IfBlock — the nested IfBlock with `block.elseif == true`
-            // owns the `} else if (EXPR){` rewrite (see branch above).
-            // Process the elseif block (which will handle its own
-            // `} else if(...) {` rewrite).
-            process_fragment_inplace(alternate, source, options, str, counter, depth);
-
-            // No closing `}` needed since the inner if block handles `{/if}`
-        } else {
-            // Find where the consequent content ends. For an empty consequent
-            // this is the body-open position (right after `{#if EXPR}`), NOT
-            // `block.start` — otherwise the `} else {` overwrite would clobber
-            // the `if(EXPR){` header we just emitted.
-            let consequent_end = if !block.consequent.nodes.is_empty() {
-                block.consequent.nodes.last().unwrap().end()
-            } else {
-                consequent_start
-            };
-
-            // For an empty `{:else}` body, the else block opens right after the
-            // `}` that closes the `{:else}` tag — NOT at `block.end` (after
-            // `{/if}`), which would make the `} else {` overwrite swallow the
-            // `{/if}` and leave the else body unclosed.
-            let alternate_start = if !alternate.nodes.is_empty() {
-                alternate_start
-            } else {
-                let bytes = source.as_bytes();
-                let mut p = consequent_end as usize;
-                while p < bytes.len() && bytes[p] != b'}' {
-                    p += 1;
-                }
-                ((p + 1).min(bytes.len())) as u32
-            };
-
-            // Overwrite {:else} with `} else {`
-            str.overwrite(consequent_end, alternate_start, "} else {");
-
-            // Hoist alternate-branch snippets above sibling declarations too.
-            hoist_snippet_blocks(alternate, source, str);
-            // Process alternate children
-            process_fragment_inplace(alternate, source, options, str, counter, depth);
-
-            // Overwrite `{/if}` with `}`
-            let alternate_end = if !alternate.nodes.is_empty() {
-                alternate.nodes.last().unwrap().end()
-            } else {
-                alternate_start
-            };
-            if alternate_end < block.end {
-                str.overwrite(alternate_end, block.end, "}");
-            }
-        }
-    } else {
-        // No alternate - just close with `}`
-        let consequent_end = if !block.consequent.nodes.is_empty() {
-            block.consequent.nodes.last().unwrap().end()
-        } else {
-            consequent_start
-        };
-        if consequent_end < block.end {
-            str.overwrite(consequent_end, block.end, "}");
-        }
-    }
-}
-
-/// Header lead-in for the each-block when CTX is relocated. Mirrors the
-/// simple-case ` for(let ` prefix; the trailing space lets the moved CTX
-/// chunk slot in cleanly.
-fn prefix_with_for(prefix: &str) -> String {
-    format!("{}for(let ", prefix)
-}
-
-/// Build the text emitted after EXPR (and the relocated CTX) in the
-/// structured-bake each-block header. Mirrors the non-relocated
-/// `header_after_expr`: `))` closes `__sveltets_2_ensureArray(EXPR)` and
-/// the `for(...)` argument list; `{` opens the for body; the idx / key
-/// bindings still travel as plain text — only CTX is source-preserved.
-fn build_each_after_ctx_tail(block: &EachBlock, source: &str) -> String {
-    let suffix = if block.context.is_some() {
-        ""
-    } else {
-        "$$each_item;"
-    };
-    // `))` closes `__sveltets_2_ensureArray(EXPR)` + the `for(...)`
-    // argument list; `{` opens the for body.
-    let mut s = format!(")){{{}", suffix);
-    if let Some(ref index) = block.index {
-        let _ = write!(s, "let {} = 1;", index);
-    }
-    if let Some(ref key) = block.key {
-        let key_text = get_expression_text(key, source);
-        s.push_str(key_text);
-        s.push(';');
-    }
-    s
-}
-
-/// Handle an each block: `{#each items as item, i (key)}...{:else}...{/each}`.
-///
-/// Generates: `for(let item of __sveltets_2_ensureArray(items)){let i = 1;key;...}`
-/// Find the byte offset of the last whitespace-bounded `as` keyword in `s`.
-fn rfind_as_keyword(s: &str) -> Option<usize> {
-    let bytes = s.as_bytes();
-    let mut found = None;
-    let mut j = 0usize;
-    while j + 1 < bytes.len() {
-        if bytes[j] == b'a' && bytes[j + 1] == b's' {
-            let before_ok = j == 0 || bytes[j - 1].is_ascii_whitespace();
-            let after_ok = bytes.get(j + 2).is_none_or(|c| c.is_ascii_whitespace());
-            if before_ok && after_ok {
-                found = Some(j);
-            }
-        }
-        j += 1;
-    }
-    found
-}
-
-/// Extend the each-collection expression's end past a trailing TypeScript
-/// postfix (`as const`, `as T`, `satisfies T`, `!`) that `remove_typescript_nodes`
-/// stripped from `block.expression`'s span. The collection is everything in the
-/// source before the each binding's ` as ` keyword (the one immediately preceding
-/// `block.context`); the parser's narrowed `expr_end` drops a trailing postfix,
-/// which official svelte2tsx keeps (e.g. `{#each link.sections! as s}` →
-/// `__sveltets_2_ensureArray(link.sections!)`). Only applies when there is a
-/// context binding (`as X`); index/key-only forms keep the narrowed end.
-fn each_collection_extended_end(block: &EachBlock, source: &str, expr_end: u32) -> u32 {
-    let Some(ctx) = block.context.as_ref() else {
-        return expr_end;
-    };
-    let Some((ctx_start, _)) = get_expression_range(ctx) else {
-        return expr_end;
-    };
-    if ctx_start <= expr_end || ctx_start as usize > source.len() {
-        return expr_end;
-    }
-    let region = slice_src(source, expr_end as usize, ctx_start as usize);
-    // The each separator is the LAST whitespace-bounded `as` before the context;
-    // everything before it (after expr_end) is the TS postfix, if any.
-    let Some(as_off) = rfind_as_keyword(region) else {
-        return expr_end;
-    };
-    let postfix = region[..as_off].trim_end();
-    // Only extend for a genuine TS postfix (`as …`, `satisfies …`, `!`). A bare
-    // `)` here is the closing paren of a `(expr)` whose wrapping parens the
-    // parser stripped symmetrically (`{#each (c) as x}`) — that must stay
-    // dropped, like the expression-tag handler. (Mirrors `handle_expression_tag`.)
-    let pf = postfix.trim_start();
-    let is_ts_postfix =
-        pf.starts_with("as ") || pf.starts_with("satisfies ") || pf.starts_with('!');
-    if !is_ts_postfix {
-        return expr_end;
-    }
-    expr_end + postfix.len() as u32
-}
-
-fn handle_each_block(
-    block: &EachBlock,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    if block.start >= block.end {
-        return;
-    }
-
-    // Expression range, extended to include a trailing TS postfix the parser
-    // narrowed away (`x!`, `x as const`).
-    let expr_range = get_expression_range(&block.expression)
-        .map(|(s, e)| (s, each_collection_extended_end(block, source, e)));
-    let expr_text = match expr_range {
-        Some((s, e)) => source.get(s as usize..e as usize).unwrap_or(""),
-        None => "",
-    };
-    let has_context = block.context.is_some();
-    let context_text = block
-        .context
-        .as_ref()
-        .map(|c| get_expression_text(c, source).to_string())
-        .unwrap_or_else(|| "$$each_item".to_string());
-
-    let body_start = if !block.body.nodes.is_empty() {
-        block.body.nodes[0].start()
-    } else {
-        block.end
-    };
-
-    // Build the for loop header.
-    // The `{#` prefix of `{#each` is replaced with spaces to preserve
-    // source positions (matching JS svelte2tsx behavior).
-    //
-    // When the loop variable shadows the collection variable (e.g., `{#each items as items}`),
-    // a temporary variable is used to avoid the shadowing issue:
-    //   `{ const $$_each = __sveltets_2_ensureArray(items); for(let items of $$_each){`
-    // Match the JS reference's prefix-spacing for `{#each ... }` headers.
-    // The JS port uses MagicString.transform() with per-position chunk moves
-    // and appendLefts; the surviving leading whitespace ends up being:
-    //   - 1 space when there's no context binding (no `as item`)
-    //   - 2 spaces when there's a context binding (`as item`)
-    //   - 3 spaces when there's a context + index binding (`as item, i`)
-    //   - 4 spaces when there's a context + index + key binding
-    //     (`as item, i (key)`)
-    // Replicate that spacing here so the column-position assertions in the
-    // language-tools fixtures match.
-    let needs_temp_var = context_text == expr_text;
-    let prefix_spaces = 1
-        + (has_context as usize)
-        + (block.index.is_some() as usize)
-        + (block.key.is_some() as usize);
-    let prefix = " ".repeat(prefix_spaces);
-
-    // Build the wrapper around the expression chunk so MagicString can
-    // preserve the expression's per-character mapping back to the
-    // original source. Context/index/key bindings come AFTER the
-    // expression in source but appear earlier (or later) in the for-loop
-    // header — bake them as ordinary text. Their column mapping is lost
-    // but they're rarely the target of type errors.
-    let (header_before_expr, header_after_expr) = if needs_temp_var {
-        (
-            format!("{}{{ const $$_each = __sveltets_2_ensureArray(", prefix),
-            {
-                let mut s = format!("); for(let {} of $$_each){{", context_text);
-                if let Some(ref index) = block.index {
-                    let _ = write!(s, "let {} = 1;", index);
-                }
-                if let Some(ref key) = block.key {
-                    let key_text = get_expression_text(key, source);
-                    s.push_str(key_text);
-                    s.push(';');
-                }
-                s
-            },
-        )
-    } else {
-        let suffix = if has_context { "" } else { "$$each_item;" };
-        (
-            format!(
-                "{}for(let {} of __sveltets_2_ensureArray(",
-                prefix, context_text
-            ),
-            {
-                let mut s = format!(")){{{}", suffix);
-                if let Some(ref index) = block.index {
-                    let _ = write!(s, "let {} = 1;", index);
-                }
-                if let Some(ref key) = block.key {
-                    let key_text = get_expression_text(key, source);
-                    s.push_str(key_text);
-                    s.push(';');
-                }
-                s
-            },
-        )
-    };
-
-    if let Some((expr_start, expr_end)) = expr_range {
-        // Try to also preserve the context binding's source range so a
-        // diagnostic on a destructuring pattern like `{ name, age }` keeps
-        // its exact column. The relocation pattern mirrors the
-        // await-with-pending case (`MagicString::move_range` + surrounding
-        // overwrites).
-        //
-        // Bails to the simpler EXPR-only preservation when:
-        //   - the context isn't an identifier or pattern with a stable
-        //     source range,
-        //   - the expression and context source ranges overlap (parser
-        //     edge case),
-        //   - the variable name collides with the expression text
-        //     (`{#each items as items}`) — the `needs_temp_var` branch
-        //     above rebakes the wrapper around the expression and would
-        //     repeat the context text twice.
-        let context_range = block.context.as_ref().and_then(get_expression_range);
-        if let (Some((ctx_s, ctx_e)), false) = (context_range, needs_temp_var)
-            && ctx_s > expr_end
-            && ctx_e <= body_start
-        {
-            // Generated header rewritten to flow as:
-            //   "  for(let " + CTX + " of __sveltets_2_ensureArray(" + EXPR + "){...rest..."
-            //
-            // We move CTX in the chunk list to before EXPR, then overwrite
-            // each surrounding gap. Idx / key bindings stay baked into
-            // the "after-expr" tail as plain text — preserving their
-            // columns would require additional relocations and offers
-            // little user value for trivial identifiers.
-            str.move_range(ctx_s, ctx_e, expr_start);
-            str.overwrite(block.start, expr_start, &prefix_with_for(&prefix));
-            str.prepend_right(expr_start, " of __sveltets_2_ensureArray(");
-            // " as " (or whitespace) between EXPR and CTX → "){...tail".
-            // Then the trailing characters between CTX and body get
-            // emitted/cleared.
-            let tail = build_each_after_ctx_tail(block, source);
-            if expr_end < ctx_s {
-                str.overwrite(expr_end, ctx_s, &tail);
-            } else {
-                str.append_left(ctx_s, &tail);
-            }
-            if ctx_e < body_start {
-                str.overwrite(ctx_e, body_start, "");
-            }
-        } else {
-            str.overwrite(block.start, expr_start, &header_before_expr);
-            if expr_end < body_start {
-                str.overwrite(expr_end, body_start, &header_after_expr);
-            } else {
-                // expr_end >= body_start (no space between expr and body opener).
-                // Append the suffix immediately after the expression chunk so
-                // MagicString anchors it at the right boundary.
-                str.append_left(expr_end, &header_after_expr);
-            }
-        }
-    } else {
-        // Parser produced no span for the expression — fall back to the
-        // monolithic bake (original behaviour).
-        let header = format!("{}{}{}", header_before_expr, expr_text, header_after_expr);
-        str.overwrite(block.start, body_start, &header);
-    }
-
-    // Hoist inner snippets to the top of the each body before processing, so
-    // their generated `const foo = ...` declarations precede the `{const}` /
-    // `{let}` declaration tags and elements that reference them.
-    hoist_snippet_blocks(&block.body, source, str);
-
-    // Process body children (each blocks don't increment depth)
-    process_fragment_inplace(&block.body, source, options, str, counter, depth);
-
-    // Handle fallback ({:else}...{/each})
-    let body_end = if !block.body.nodes.is_empty() {
-        block.body.nodes.last().unwrap().end()
-    } else {
-        body_start
-    };
-
-    if let Some(ref fallback) = block.fallback {
-        let fallback_start = if !fallback.nodes.is_empty() {
-            fallback.nodes[0].start()
-        } else {
-            block.end
-        };
-
-        // Overwrite {:else} with `}`
-        str.overwrite(body_end, fallback_start, "}");
-
-        // Process fallback
-        process_fragment_inplace(fallback, source, options, str, counter, depth);
-
-        let fallback_end = if !fallback.nodes.is_empty() {
-            fallback.nodes.last().unwrap().end()
-        } else {
-            fallback_start
-        };
-
-        if fallback_end < block.end {
-            str.overwrite(fallback_end, block.end, "");
-        }
-    } else {
-        // Close the for loop
-        let closing = if needs_temp_var { "}}" } else { "}" };
-        if body_end < block.end {
-            str.overwrite(body_end, block.end, closing);
-        } else {
-            // Empty each body (`{#each x as i}{/each}`): body_end == block.end,
-            // so there is no source region left to overwrite with the closing
-            // brace (the opening-tag remainder + `{/each}` were already cleared
-            // by the header handling). Append it so the `for(...){` opened by
-            // the header is balanced — otherwise the unclosed brace cascades up
-            // and leaves `$$render` itself unterminated.
-            str.append_left(block.end, closing);
-        }
-    }
-}
-
-/// Handle an await block: `{#await promise}...{:then value}...{:catch error}...{/await}`.
-///
-/// Generates patterns like:
-/// - `{#await promise}pending{:then value}resolved{/await}`
-///   → `{  { const $$_value = await (promise);{ const value = $$_value; resolved}}}`
-/// - `{#await promise then value}resolved{/await}`
-///   → `{  { const $$_value = await (promise);{ const value = $$_value; resolved}}`
-/// - `{#await promise catch error}rejected{/await}`
-///   → `{  { try { const $$_value = await (promise);} catch(error) { rejected}}`
-fn handle_await_block(
-    block: &AwaitBlock,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    if block.start >= block.end {
-        return;
-    }
-
-    let expr_text = get_expression_text(&block.expression, source);
-
-    // Determine the structure of the await block:
-    // 1. `{#await promise}` pending `{:then value}` then `{/await}` (has pending, then)
-    // 2. `{#await promise then value}` then `{/await}` (no pending, immediate then)
-    // 3. `{#await promise catch error}` catch `{/await}` (no pending, immediate catch)
-    // 4. `{#await promise}` pending `{:then value}` then `{:catch error}` catch `{/await}`
-
-    let has_pending = block.pending.as_ref().is_some_and(|p| !p.nodes.is_empty());
-    let has_then = block.then.is_some();
-    let has_catch = block.catch.is_some();
-
-    let value_text = block
-        .value
-        .as_ref()
-        .map(|v| get_expression_text(v, source).to_string())
-        .unwrap_or_default();
-
-    let error_text = block
-        .error
-        .as_ref()
-        .map(|e| get_expression_text(e, source).to_string())
-        .unwrap_or_default();
-
-    if has_pending {
-        // Pattern: {#await promise} pending {:then value} then {:catch error} catch {/await}
-        let pending = block.pending.as_ref().unwrap();
-        let pending_start = if !pending.nodes.is_empty() {
-            pending.nodes[0].start()
-        } else {
-            block.end
-        };
-
-        // Handle then
-        if let Some(ref then) = block.then {
-            let then_start = if !then.nodes.is_empty() {
-                then.nodes[0].start()
-            } else {
-                block.end
-            };
-
-            let prev_end = if !pending.nodes.is_empty() {
-                pending.nodes.last().unwrap().end()
-            } else {
-                pending_start
-            };
-
-            // The PROMISE expression source-wise lives inside the
-            // `{#await PROMISE}` opener but generated-wise belongs at the
-            // `{:then VALUE}` boundary. `move_range` relocates the
-            // expression chunk past the pending fragment so its
-            // per-character source map survives intact; the `const
-            // $$_value = await (…); { const VALUE = $$_value; ` wrapper
-            // is attached as the relocated chunk's intro / outro so it
-            // travels with the expression.
-            if let Some((expr_start, expr_end)) = get_expression_range(&block.expression) {
-                str.move_range(expr_start, expr_end, prev_end);
-                str.overwrite(block.start, expr_start, "   { ");
-                if expr_end < pending_start {
-                    str.overwrite(expr_end, pending_start, "");
-                }
-                // When a `catch` (or error variable) is present, the await
-                // must be wrapped in a `try {` so the later `} catch(...) {`
-                // is balanced. Mirrors upstream `handleAwait` emitting
-                // `try { ` whenever `error || !catch.skip`.
-                // `const $$_value = ` and the `{ const VALUE = $$_value; ` inner
-                // block are emitted ONLY when there's a `{:then value}` binding
-                // (mirrors official `handleAwait`, which gates both on
-                // `awaitBlock.value`). A bare `{:then}` is just `await (…);` with
-                // the then-body inline.
-                str.prepend_right(
-                    expr_start,
-                    match (has_catch, value_text.is_empty()) {
-                        (true, false) => "try { const $$_value = await (",
-                        (true, true) => "try { await (",
-                        (false, false) => "const $$_value = await (",
-                        (false, true) => "await (",
-                    },
-                );
-                let suffix = if !value_text.is_empty() {
-                    format!(");{{ const {} = $$_value; ", value_text)
-                } else {
-                    ");".to_string()
-                };
-                str.append_left(expr_end, &suffix);
-                if prev_end < then_start {
-                    str.overwrite(prev_end, then_start, "");
-                }
-                process_fragment_inplace(pending, source, options, str, counter, depth);
-            } else {
-                // Parser couldn't span the expression — fall back to
-                // the original monolithic bake.
-                str.overwrite(block.start, pending_start, "   { ");
-                process_fragment_inplace(pending, source, options, str, counter, depth);
-                // `try { ` wrapper when a catch/error is present (see above).
-                let try_prefix = if has_catch { "try { " } else { "" };
-                if !value_text.is_empty() {
-                    str.overwrite(
-                        prev_end,
-                        then_start,
-                        &format!(
-                            "{}const $$_value = await ({});{{ const {} = $$_value; ",
-                            try_prefix, expr_text, value_text
-                        ),
-                    );
-                } else {
-                    str.overwrite(
-                        prev_end,
-                        then_start,
-                        &format!("{}const $$_value = await ({});{{ ", try_prefix, expr_text),
-                    );
-                }
-            }
-
-            process_fragment_inplace(then, source, options, str, counter, depth);
-
-            // Handle catch after then
-            if let Some(ref catch) = block.catch {
-                let catch_start = if !catch.nodes.is_empty() {
-                    catch.nodes[0].start()
-                } else {
-                    block.end
-                };
-
-                let then_end = if !then.nodes.is_empty() {
-                    then.nodes.last().unwrap().end()
-                } else {
-                    then_start
-                };
-
-                // Close the `try` (always) plus the value block (only when a
-                // `{:then value}` binding opened one), then open the catch.
-                let close_before_catch = if value_text.is_empty() { "}" } else { "}}" };
-                if !error_text.is_empty() {
-                    str.overwrite(
-                        then_end,
-                        catch_start,
-                        &format!(
-                            "{} catch($$_e) {{ const {} = __sveltets_2_any();",
-                            close_before_catch, error_text
-                        ),
-                    );
-                } else {
-                    str.overwrite(
-                        then_end,
-                        catch_start,
-                        &format!("{} catch($$_e) {{ ", close_before_catch),
-                    );
-                }
-
-                process_fragment_inplace(catch, source, options, str, counter, depth);
-
-                let catch_end = if !catch.nodes.is_empty() {
-                    catch.nodes.last().unwrap().end()
-                } else {
-                    catch_start
-                };
-
-                if catch_end < block.end {
-                    str.overwrite(catch_end, block.end, "}}");
-                }
-            } else {
-                // No catch: close the value block (if any) + the outer await
-                // block. A bare `{:then}` opened only the outer block.
-                let then_end = if !then.nodes.is_empty() {
-                    then.nodes.last().unwrap().end()
-                } else {
-                    then_start
-                };
-                if then_end < block.end {
-                    let close = if value_text.is_empty() { "}" } else { "}}" };
-                    str.overwrite(then_end, block.end, close);
-                }
-            }
-        } else {
-            // No `:then` after the pending block. Covers
-            // `{#await p}pending{/await}` (pending only) and
-            // `{#await p}pending{:catch e}…{/await}` (pending + catch, no then).
-            // Previously this branch emitted only a trailing `}` — it never
-            // opened the block, dropped the `await(promise)` entirely, and
-            // ignored the catch, producing brace-unbalanced / invalid TSX.
-            // Mirror upstream `handleAwait`: `{ <pending> [try {] await(p);
-            // [} catch($$_e) { … }] }`.
-            let pending_end = if !pending.nodes.is_empty() {
-                pending.nodes.last().unwrap().end()
-            } else {
-                pending_start
-            };
-
-            // Opening `{ ` — consume the `{#await PROMISE}` opener (PROMISE is
-            // re-emitted as `await(...)` after the pending body).
-            str.overwrite(block.start, pending_start, "   { ");
-            process_fragment_inplace(pending, source, options, str, counter, depth);
-
-            if let Some(ref catch) = block.catch {
-                let catch_start = if !catch.nodes.is_empty() {
-                    catch.nodes[0].start()
-                } else {
-                    block.end
-                };
-                let header = if !error_text.is_empty() {
-                    format!(
-                        "try {{ await ({});}} catch($$_e) {{ const {} = __sveltets_2_any();",
-                        expr_text, error_text
-                    )
-                } else {
-                    format!("try {{ await ({});}} catch($$_e) {{ ", expr_text)
-                };
-                if pending_end < catch_start {
-                    str.overwrite(pending_end, catch_start, &header);
-                } else {
-                    str.append_left(pending_end, &header);
-                }
-                process_fragment_inplace(catch, source, options, str, counter, depth);
-                let catch_end = if !catch.nodes.is_empty() {
-                    catch.nodes.last().unwrap().end()
-                } else {
-                    catch_start
-                };
-                if catch_end < block.end {
-                    str.overwrite(catch_end, block.end, "}}");
-                }
-            } else if pending_end < block.end {
-                str.overwrite(pending_end, block.end, &format!("await ({});}}", expr_text));
-            }
-        }
-    } else if has_then {
-        // Pattern: {#await promise then value} then {/await} (no pending)
-        // Or:      {#await promise then value} then {:catch error} catch {/await}
-        let then = block.then.as_ref().unwrap();
-        let then_start = if !then.nodes.is_empty() {
-            then.nodes[0].start()
-        } else {
-            block.end
-        };
-
-        // In source order, `{#await PROMISE then VALUE}` is followed
-        // directly by the then-body. The generated wrapper also places
-        // the expression before VALUE (and VALUE before the body), so
-        // we can preserve PROMISE's chunk in place by splitting the
-        // header overwrite into a prefix / suffix pair around the
-        // expression range.
-        // `const $$_value = ` and the `{ const VALUE = $$_value; … }` scope are
-        // emitted only for a `{:then value}` binding (mirrors official
-        // `handleAwait`, which gates both on `awaitBlock.value`). A bare
-        // `{#await … then}` is just `await (…);` with the body inline (the body
-        // elements provide their own block). `value_close` is the matching `}`
-        // for the value scope, emitted by the close logic below.
-        let value_close = if value_text.is_empty() { "" } else { "}" };
-        let (header_prefix, header_suffix) = if has_catch {
-            (
-                if value_text.is_empty() {
-                    "   { try { await ("
-                } else {
-                    "   { try { const $$_value = await ("
-                },
-                if !value_text.is_empty() {
-                    format!(");{{ const {} = $$_value; ", value_text)
-                } else {
-                    ");".to_string()
-                },
-            )
-        } else {
-            (
-                if value_text.is_empty() {
-                    "   { await ("
-                } else {
-                    "   { const $$_value = await ("
-                },
-                if !value_text.is_empty() {
-                    format!(");{{ const {} = $$_value; ", value_text)
-                } else {
-                    ");".to_string()
-                },
-            )
-        };
-
-        if let Some((expr_start, expr_end)) = get_expression_range(&block.expression) {
-            str.overwrite(block.start, expr_start, header_prefix);
-            if expr_end < then_start {
-                str.overwrite(expr_end, then_start, &header_suffix);
-            } else {
-                str.append_left(expr_end, &header_suffix);
-            }
-        } else {
-            str.overwrite(
-                block.start,
-                then_start,
-                &format!("{}{}{}", header_prefix, expr_text, header_suffix),
-            );
-        }
-
-        process_fragment_inplace(then, source, options, str, counter, depth);
-
-        let then_end = if !then.nodes.is_empty() {
-            then.nodes.last().unwrap().end()
-        } else {
-            then_start
-        };
-
-        if has_catch {
-            // Handle catch after then
-            let catch = block.catch.as_ref().unwrap();
-            let catch_start = if !catch.nodes.is_empty() {
-                catch.nodes[0].start()
-            } else {
-                block.end
-            };
-
-            if !error_text.is_empty() {
-                str.overwrite(
-                    then_end,
-                    catch_start,
-                    &format!(
-                        "{}}} catch($$_e) {{ const {} = __sveltets_2_any();",
-                        value_close, error_text
-                    ),
-                );
-            } else {
-                // Close the value block (only when there's a `{:then value}`
-                // binding) + `try`, then open the catch. Always emit `($$_e)`.
-                str.overwrite(
-                    then_end,
-                    catch_start,
-                    &format!("{}}} catch($$_e) {{ ", value_close),
-                );
-            }
-
-            process_fragment_inplace(catch, source, options, str, counter, depth);
-
-            let catch_end = if !catch.nodes.is_empty() {
-                catch.nodes.last().unwrap().end()
-            } else {
-                catch_start
-            };
-
-            if catch_end < block.end {
-                str.overwrite(catch_end, block.end, "}}");
-            }
-        } else {
-            // Close the value block (if any) + the outer await block. This
-            // handles both the normal case (then_end < block.end: the then
-            // body ends before {/await}, so we overwrite the gap) and the
-            // empty-then-body case (then_end == block.end: the overwrite from
-            // expr_end to block.end already consumed that region, so we must
-            // append rather than overwrite a zero-length range).
-            let close = format!("{}}}", value_close);
-            if then_end < block.end {
-                str.overwrite(then_end, block.end, &close);
-            } else {
-                str.append_left(block.end, &close);
-            }
-        }
-    } else if has_catch {
-        // Pattern: {#await promise catch error} catch {/await} (no pending, no then)
-        let catch = block.catch.as_ref().unwrap();
-        let catch_start = if !catch.nodes.is_empty() {
-            catch.nodes[0].start()
-        } else {
-            block.end
-        };
-
-        let (header_prefix, header_suffix) = (
-            "   { try { await (",
-            if !error_text.is_empty() {
-                format!(
-                    ");}} catch($$_e) {{ const {} = __sveltets_2_any();",
-                    error_text
-                )
-            } else {
-                ");} catch($$_e) { ".to_string()
-            },
-        );
-        if let Some((expr_start, expr_end)) = get_expression_range(&block.expression) {
-            str.overwrite(block.start, expr_start, header_prefix);
-            if expr_end < catch_start {
-                str.overwrite(expr_end, catch_start, &header_suffix);
-            } else {
-                str.append_left(expr_end, &header_suffix);
-            }
-        } else if !error_text.is_empty() {
-            str.overwrite(
-                block.start,
-                catch_start,
-                &format!(
-                    "   {{ try {{ await ({});}} catch($$_e) {{ const {} = __sveltets_2_any();",
-                    expr_text, error_text
-                ),
-            );
-        } else {
-            str.overwrite(
-                block.start,
-                catch_start,
-                &format!("   {{ try {{ await ({});}} catch($$_e) {{ ", expr_text),
-            );
-        }
-
-        process_fragment_inplace(catch, source, options, str, counter, depth);
-
-        let catch_end = if !catch.nodes.is_empty() {
-            catch.nodes.last().unwrap().end()
-        } else {
-            catch_start
-        };
-
-        if catch_end < block.end {
-            str.overwrite(catch_end, block.end, "}}");
-        }
-    } else {
-        // Bare await block `{#await promise}{/await}` (no pending/then/catch).
-        // Official `handleAwait` emits `{ await (EXPR);}` — the promise is
-        // always awaited, so the `await` keyword must be present (it was
-        // previously dropped, emitting `{EXPR;}`).
-        if let Some((expr_start, expr_end)) = get_expression_range(&block.expression) {
-            str.overwrite(block.start, expr_start, "{ await (");
-            if expr_end < block.end {
-                str.overwrite(expr_end, block.end, ");}");
-            } else {
-                str.append_left(expr_end, ");}");
-            }
-        } else {
-            str.overwrite(
-                block.start,
-                block.end,
-                &format!("{{ await ({});}}", expr_text),
-            );
-        }
-    }
-}
-
-/// Handle a key block: `{#key expression}...{/key}`.
-fn handle_key_block(
-    block: &KeyBlock,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    if block.start >= block.end {
-        return;
-    }
-
-    let expr_text = get_expression_text(&block.expression, source);
-
-    // For an empty `{#key EXPR}{/key}` body, the block scope opens right after
-    // the `}` that closes the `{#key EXPR}` tag — NOT at `block.end` (after
-    // `{/key}`), which would make the header rewrite swallow `{/key}` and leave
-    // the `{` unbalanced.
-    let content_start = if !block.fragment.nodes.is_empty() {
-        block.fragment.nodes[0].start()
-    } else {
-        let expr_end = get_expression_range(&block.expression)
-            .map(|(_, e)| e)
-            .unwrap_or(block.start);
-        let bytes = source.as_bytes();
-        let mut p = expr_end as usize;
-        while p < bytes.len() && bytes[p] != b'}' {
-            p += 1;
-        }
-        ((p + 1).min(bytes.len())) as u32
-    };
-
-    // Preserve the expression chunk in place so its per-character mapping
-    // survives. Official emits the key expression as a bare statement followed
-    // by a block scope for the body — `{#key value}…{/key}` → `value; { … }`
-    // (NOT `{ value; … }`). So drop the `{#key ` prefix and turn the closing
-    // `}` of the opening tag into `; {`. Mirrors KeyBlock handling in upstream
-    // htmlxtojsx_v2.
-    if let Some((expr_start, expr_end)) = get_expression_range(&block.expression) {
-        str.overwrite(block.start, expr_start, "");
-        if expr_end < content_start {
-            str.overwrite(expr_end, content_start, "; {");
-        } else {
-            str.append_left(expr_end, "; {");
-        }
-    } else {
-        str.overwrite(block.start, content_start, &format!("{expr_text}; {{"));
-    }
-
-    // Process children
-    process_fragment_inplace(&block.fragment, source, options, str, counter, depth);
-
-    let content_end = if !block.fragment.nodes.is_empty() {
-        block.fragment.nodes.last().unwrap().end()
-    } else {
-        content_start
-    };
-
-    if content_end < block.end {
-        str.overwrite(content_end, block.end, "}");
-    }
-}
-
-/// Handle a snippet block: `{#snippet name(params)}...{/snippet}`.
-///
-/// Generates:
-/// ```text
-/// const name = (params): ReturnType<import('svelte').Snippet> => { async () => {
-///   ...
-/// };return __sveltets_2_any(0)};
-/// ```
-fn handle_snippet_block(
-    block: &SnippetBlock,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    handle_snippet_block_inner(block, source, options, str, counter, false, depth);
-}
-
-/// Transform a `{#snippet name(params)}` block that is a direct child of a
-/// component into an **implicit prop**: `name:(params) => { async () => { …body…
-/// };return __sveltets_2_any(0)},`. Unlike the standalone form there is no
-/// leading `const`, no `: ReturnType<…>` annotation, and the closing ends in a
-/// `,` so the result drops straight into the component's `props: { … }` object
-/// literal (the caller relocates the range there via `move_range`). This mirrors
-/// upstream svelte2tsx `addImplicitSnippetProp`, and lets TypeScript
-/// contextually type the snippet's parameters from the prop's `Snippet<[T]>`
-/// type while satisfying required snippet props (#780).
-fn handle_snippet_block_as_component_prop(
-    block: &SnippetBlock,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    depth: u32,
-) {
-    handle_snippet_block_inner(block, source, options, str, counter, true, depth);
-}
-
-fn handle_snippet_block_inner(
-    block: &SnippetBlock,
-    source: &str,
-    options: &Svelte2TsxOptions,
-    str: &mut MagicString,
-    counter: &mut Counter,
-    as_component_prop: bool,
-    // Snippet bodies always start at depth 0 (official resets `element` on
-    // entry), so the inherited depth is intentionally unused.
-    _depth: u32,
-) {
-    if block.start >= block.end {
-        return;
-    }
-
-    let name_text = get_expression_text(&block.expression, source);
-
-    // Build parameters string
-    let params_text = if !block.parameters.is_empty() {
-        block
-            .parameters
-            .iter()
-            .map(|p| get_expression_text(p, source))
-            .collect::<Vec<_>>()
-            .join(", ")
-    } else {
-        String::new()
-    };
-
-    let has_body_nodes = !block.body.nodes.is_empty();
-    let body_start = if has_body_nodes {
-        block.body.nodes[0].start()
-    } else {
-        block.end
-    };
-
-    // Overwrite `{#snippet name(params)}` with function declaration.
-    // Position markers are added to help the language server:
-    // - `/*Ωignore_positionΩ*/` after the name and after `async ()`
-    // - Return type wrapped in `/*Ωignore_startΩ*/.../*Ωignore_endΩ*/`
-    //
-    // Two emission modes match the JS reference (`SnippetBlock.ts`):
-    // - TS syntax (TS file or non-JSDoc emit): `: ReturnType<...>` after the
-    //   parameter list, with `<typeParams>` if the snippet declared generics
-    // - JSDoc syntax (JS file + JSDoc emit): `/** @returns {ReturnType<...>} */`
-    //   before the `(params)` arrow, no generic-params syntax
-    let use_ts_syntax = options.is_ts_file || !options.emit_jsdoc;
-    let type_params_str = match (use_ts_syntax, block.type_params.as_ref()) {
-        (true, Some(tp)) => format!("<{}>", tp),
-        _ => String::new(),
-    };
-    // Implicit-prop form (`name:(params) => …`) vs standalone declaration
-    // (`const name = (params): ReturnType<…> => …`). The implicit form omits the
-    // leading `const`, the return-type annotation, and the generic `<typeParams>`
-    // — mirroring upstream's `addImplicitSnippetProp` transforms — and closes
-    // with a trailing `,` so it slots into the component `props` object literal.
-    let header = if as_component_prop {
-        format!(
-            "{}:({}) => {{ async ()/*\u{03A9}ignore_position\u{03A9}*/ => {{",
-            name_text, params_text
-        )
-    } else if use_ts_syntax {
-        // Single leading space (the overwrite replaces `{#snippet ` whose leading
-        // `{` becomes the space) — matches official; oxfmt normalises it for
-        // valid output, but a top-level-await component is emitted raw.
-        format!(
-            " const {}/*\u{03A9}ignore_position\u{03A9}*/ = {}({})/*\u{03A9}ignore_start\u{03A9}*/: ReturnType<import('svelte').Snippet>/*\u{03A9}ignore_end\u{03A9}*/ => {{ async ()/*\u{03A9}ignore_position\u{03A9}*/ => {{",
-            name_text, type_params_str, params_text
-        )
-    } else {
-        // JSDoc emission uses one fewer leading space (the `/** @returns */`
-        // marker takes the visual slot otherwise occupied by the TS `:` and
-        // its surrounding `/*Ωignore*/` comments).
-        format!(
-            " const {}/*\u{03A9}ignore_position\u{03A9}*/ = /** @returns {{ReturnType<import('svelte').Snippet>}} */ ({}) => {{ async ()/*\u{03A9}ignore_position\u{03A9}*/ => {{",
-            name_text, params_text
-        )
-    };
-    let closing = if as_component_prop {
-        "};return __sveltets_2_any(0)},"
-    } else {
-        "};return __sveltets_2_any(0)};"
-    };
-    if has_body_nodes {
-        str.overwrite(block.start, body_start, &header);
-        // Process body at depth 0: official resets `element = undefined` when
-        // entering a SnippetBlock, so element/component names inside a snippet
-        // body always count depth from the snippet (e.g. `<Child>` directly in
-        // a snippet is `$$_…C0C`), regardless of how deeply the snippet itself
-        // is nested in elements / `<svelte:boundary>`.
-        process_fragment_inplace(&block.body, source, options, str, counter, 0);
-
-        let body_end = block.body.nodes.last().unwrap().end();
-        if body_end < block.end {
-            // Overwrite `{/snippet}` with closing
-            str.overwrite(body_end, block.end, closing);
-        }
-    } else {
-        // Empty body: collapse the whole `{#snippet name(params)}{/snippet}`
-        // into a single declaration. Without this branch the closing
-        // `};return __sveltets_2_any(0)};` was never emitted because both the
-        // body-start overwrite and the would-be closing overwrite landed at
-        // the same offset.
-        let combined = format!("{}{}", header, closing);
-        str.overwrite(block.start, block.end, &combined);
-    }
-}
 
 // =============================================================================
 // Element Nodes
